Move token formatting code to the main sucrase package
This should make it easier to display things like the computed token context.
