+++ src/transformers/FlowTransformer.ts
@@ -1,5 +1,5 @@
-import {RootTransformer} from "../index";
 import TokenProcessor from "../TokenProcessor";
+import RootTransformer from "./RootTransformer";
 import Transformer from "./Transformer";
 
 export default class FlowTransformer extends Transformer {
