+++ src/transformers/JSXTransformer.ts
@@ -1,6 +1,6 @@
-import {RootTransformer} from "../index";
 import TokenProcessor from "../TokenProcessor";
 import IdentifierReplacer from "./IdentifierReplacer";
+import RootTransformer from "./RootTransformer";
 import Transformer from "./Transformer";
 
 export default class JSXTransformer extends Transformer {
